requirements.txt
python-telegram-bot==21.6
supabase==2.13.0
google-generativeai==0.7.2
google-auth==2.34.0
google-api-python-client==2.147.0
Pillow==10.4.0
python-dotenv==1.0.1
